version 7.0.0 알림기능 추가 예외처리 추가

+++ src/main/java/com/example/noticeboard/service/CommentServiceImpl.java
@@ -1,21 +1,33 @@
 package com.example.noticeboard.service;
 
 import com.example.noticeboard.dto.CommentDTO;
+import com.example.noticeboard.dto.SliceRequestDTO;
+import com.example.noticeboard.dto.SliceResultDTO;
 import com.example.noticeboard.entity.Board;
 import com.example.noticeboard.entity.Comment;
+import com.example.noticeboard.entity.MemberComment;
+import com.example.noticeboard.entity.member.Member;
+import com.example.noticeboard.entity.notification.NotificationType;
+import com.example.noticeboard.exception.ErrorCode;
+import com.example.noticeboard.exception.custom_exception.AccessDeniedException;
 import com.example.noticeboard.exception.custom_exception.CommentNotFoundException;
+import com.example.noticeboard.exception.custom_exception.DuplicateException;
+import com.example.noticeboard.repository.MemberCommentRepository;
 import com.example.noticeboard.repository.commentRepository.CommentRepository;
+import com.example.noticeboard.security.dto.MemberDTO;
+import com.example.noticeboard.service.event.NotificationEvent;
+import com.example.noticeboard.service.event.RemoveBoardEvent;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
-import org.springframework.dao.EmptyResultDataAccessException;
-import org.springframework.data.domain.Sort;
+import org.springframework.context.ApplicationEventPublisher;
+import org.springframework.context.event.EventListener;
+import org.springframework.data.domain.Slice;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.util.List;
-import java.util.NoSuchElementException;
+import java.util.Map;
 import java.util.Optional;
-import java.util.stream.Collectors;
+import java.util.function.Function;
 
 @Service
 @Log4j2
@@ -23,55 +35,94 @@ import java.util.stream.Collectors;
 @Transactional
 public class CommentServiceImpl implements CommentService {
 
+
     private final CommentRepository commentRepository;
 
+    private final MemberCommentRepository memberCommentRepository;
+
+    private final ApplicationEventPublisher publisher;
+
+
+
+    //댓글 리스트 출력 함수
     @Override
-    public List<CommentDTO> getCommentList(Long boardId, String order) {
-
-        Sort sort;
-        if(order.equals("new")) sort = Sort.by("createDate").descending();
-        else if(order.equals("old")) sort  = Sort.by("createDate").ascending();
-        else if(order.equals("recomend")) sort = Sort.by("recomendNum").descending();
-        else sort = Sort.by("id").descending();
-
+    @Transactional(readOnly = true)
+    public SliceResultDTO<CommentDTO,Map<String, Object>> getCommentList(Long boardId, SliceRequestDTO sliceRequestDTO) {
         Board board = Board.builder().id(boardId).build();
-        List<Comment> commentList = commentRepository.getCommentList(board, sort);
-        return commentList.stream().map(comment -> entityToDto(comment)).collect(Collectors.toList());
+        Function<Map<String, Object>, CommentDTO> fn = (map-> entityToDto((Comment) map.get("comment"), (Long) map.get("childCommentCount")));
+        Slice<Map<String, Object>> result = commentRepository.getCommentListByBoard(board, sliceRequestDTO.getPageable());
+        return new SliceResultDTO<>(result, fn);
+    }
+
+    // 답글 리스트 출력 함수
+    @Override
+    @Transactional(readOnly = true)
+    public SliceResultDTO<CommentDTO, Comment> getChildCommentList(Long commentId, SliceRequestDTO sliceRequestDTO) {
+        Comment comment = Comment.builder().id(commentId).build();
+        Function<Comment, CommentDTO> fn = (en-> entityToDto(en,0l));
+        Slice<Comment> result = commentRepository.getCommentListByParent(comment, sliceRequestDTO.getPageable());
+        log.info(result);
+        return new SliceResultDTO<>(result, fn);
     }
 
     @Override
-    public int recommend(Long id) {
+    public Integer recommend(Long id, MemberDTO recommender) {
+        Member member = Member.builder().username(recommender.getUsername()).build();
+
         Optional<Comment> result = commentRepository.findById(id);
-        if(result.isEmpty()) throw new CommentNotFoundException("존재하지 않는 댓글입니다.");
+        if(result.isEmpty()) throw new CommentNotFoundException(ErrorCode.NOT_FOUND_COMMENT); // 댓글이 존재하지 않는다면
         Comment comment = result.get();
+
+        //자신이 쓴 댓글이라면 추천 x
+        if(comment.getMember().getUsername().equals(recommender.getUsername())) throw new AccessDeniedException(ErrorCode.SELF_RECOMMEND);;
+
+        //이미 추천한 댓글이라면 추천 x
+        if(memberCommentRepository.existsByMemberAndComment(member, comment)) throw new DuplicateException(ErrorCode.DUPLICATE_RECOMMEND);
+
+        MemberComment memberComment = MemberComment.builder().member(member).comment(comment).build();
+        memberCommentRepository.save(memberComment);
+
         comment.addRecomendNum();
+
+        //알림 이벤트 발행
+        publisher.publishEvent(new NotificationEvent(comment, recommender.getUsername(), NotificationType.COMMENT_RECOMMEND));
+
         return comment.getRecomendNum();
     }
 
     @Override
-    public Long modify(CommentDTO commentDTO) {
+    public Long modify(CommentDTO commentDTO, MemberDTO memberDTO) {
         Optional<Comment> result = commentRepository.findById(commentDTO.getId());
-        if(result.isEmpty()) throw new CommentNotFoundException("존재하지 않는 댓글입니다.");
+        if(result.isEmpty()) throw new CommentNotFoundException(ErrorCode.NOT_FOUND_COMMENT);
         Comment comment = result.get();
+        if(!memberDTO.getUsername().equals(comment.getMember().getUsername())) throw new AccessDeniedException(ErrorCode.ACCESS_DENIED);
         comment.changeContent(commentDTO.getContent());
         return comment.getId();
     }
 
     @Override
-    public Long remove(Long id) {
-        try {
-            commentRepository.deleteById(id);
-        }catch (EmptyResultDataAccessException e){
-            throw new CommentNotFoundException("존재하지 않는 댓글입니다.");
-        }
+    public Long remove(Long id, MemberDTO memberDTO) {
+        Optional<Comment> result = commentRepository.findById(id);
+        if(result.isEmpty()) throw new CommentNotFoundException(ErrorCode.NOT_FOUND_COMMENT);
+        Comment comment = result.get();
+        if(!memberDTO.getUsername().equals(comment.getMember().getUsername())) throw new AccessDeniedException(ErrorCode.ACCESS_DENIED);
+        memberCommentRepository.deleteByComment(Comment.builder().id(id).build());
+        commentRepository.deleteById(id);
         return id;
     }
 
+    @EventListener
+    public void RemoveByBoard(RemoveBoardEvent removeBoardEvent){
+        commentRepository.deleteByBoard(removeBoardEvent.getBoard());
+    }
+
     @Override
-    public Long register(CommentDTO commentDTO) {
+    public CommentDTO register(CommentDTO commentDTO) {
+
         Comment comment = dtoToEntity(commentDTO);
-        log.info(comment.toString());
         commentRepository.save(comment);
-        return comment.getId();
+        //알림 이벤트 발행
+        publisher.publishEvent(new NotificationEvent(comment, commentDTO.getReceiver(), NotificationType.COMMENT_REGISTER));
+        return entityToDto(comment, 0l);
     }
 }
